refactor(schema-api): extract LockApi trait
Extract LockApi trait from monolithic SchemaApi to improve code
organization and maintainability. This continues the ongoing schema API
refactoring effort by moving lock management operations into their
own focused trait interface.
