+++ src/meta/api/src/lib.rs
@@ -28,6 +28,7 @@ pub mod index_api;
 pub mod kv_app_error;
 pub mod kv_pb_api;
 pub mod kv_pb_crud_api;
+pub mod lock_api;
 pub mod meta_txn_error;
 pub mod name_id_value_api;
 pub mod name_value_api;
@@ -52,6 +53,7 @@ pub use data_mask_api::DatamaskApi;
 pub use database_api::DatabaseApi;
 pub use dictionary_api::DictionaryApi;
 pub use index_api::IndexApi;
+pub use lock_api::LockApi;
 pub use row_access_policy_api::RowAccessPolicyApi;
 pub use schema_api::SchemaApi;
 pub use schema_api_test_suite::SchemaApiTestSuite;
